Updated Sections 2.1, 4.1, and 4.2
All thats left is sections 2.1, 4.4, and 5. User's stories need to be included somewhere in the document. After the remaining sections are added, A final polish and proofread are needed. The final polish needs to remove all text comments in the <> and needs to clean up the overall formatting of the document.